chore: Move UserNode and GroupNode to gql_models module
